requirements.txt
flask==3.0.3
rembg==2.0.56
pillow==10.4.0
numpy==2.1.1
onnxruntime==1.17.3
